Bring Ndef push usage up to SDK level 14
The way Ndef push was enabled and disabled was not up to
recommendations anyway, it did not follow the activity lifecycle.

At Coreorient we can target Ice Cream as the baseline, and
expect devices with this or newer Android version to be widely
deployed when we ship to the general public.
Even some older devices such as Nexus S are already updated
to Jelly Bean. For pre-Ice Cream holdouts, see
http://www.nfcworld.com/nfc-phones-list/#available